6b also look for messages
Packet start marker was just four, but now a start of message is marked
by fourteen unique characters. That's as easy in source code as before,
just make the marker length a parameter.

This would be not very efficient for long messages and long markers
though, as the hashset is rebuilt from almost identical input for each
step. A better way could be to scan(), maintaining a (hash)map that is
popped and pushed for the exiting and entering characters.