7 no space left on device
Parse that odd specification of a filesystem structure from a
"command-line" session. Traversing it for the unusual directory size
where those nested sizes are counted twice is then much easier once the
tree has been set up.

This time Rust provides some significant grief with the data structure I
happened to choose, as the current directory stack needs references to
the tree structure that also is being modified, the currently modified
directory also is a reference to the already created tree structure...
As this looked like a case where Rc and RefCell would fit, they indeed
fit and something just happened I'm not sure if those weak refs are
needed.

One possibly easier alternative would be a flat list of filesystem
entries and a separate list of refs to them for the dir tree, possibly
with indices because refs would break when the flat list could
reallocate. But indices are the worst programming pattern for high level
logic, they're like weak refs that appear out of nowhere.