7b choose a directory to delete
The original tree traversal isn't that good anymore so make a more
generic one in postorder to look at sizes. Compute the total size and
the size of each directory of suitable size (this time with no nested
duplicates) and then look up the smallest of the suitable deletion
candidates.

This also got redesigned somewhat while being implemented, so it'll need
another cleanup iteration later on, but at least the result is correct.