simplify 7 for just sizes
The traversal is needed only for directory sizes. Make it so.

Traversal is still of that kludgy visit-callback-type. A generator would
be much easier to write for this, but this is not lisp. An iterator is
possible but would need an explicit traversal stack and some more
details, maybe later.

Simplify parsing: current dir is better as top of the visit stack, and
the weak refs are not needed.