7: practice tree traversal iterators
While the visit pattern is easy to see and recurse with, it needs that
clumsy outer state for size sums. Implement post-order tree traversal
with the Iterator trait mainly for fun and technical practice. This
iterator yields directory sizes from files and subdirectories within.

The iterator is implemented with an internal stack. Three types of
entries mark proceeding and completed sub-size calculation operations
and sentinels for propagating a completed size into the iterator
interface.

The stack unfortunately does mean various allocations during traversal.
Recursion involves an implicit traversal stack too, but that happens on
the thread stack which is simpler than heap reallocation.

With some further work, this could be made into a more generic tree
traversal "library"; for now the size operations are hardcoded.
Generality might be achieved cleanly e.g. by using an internal scan or
fold iterator provided as a type parameter, or just with some closures.