9b ropes whipping as you fall
Extend the simulation to support longer ropes. The movement logic is the
same for each pair of knots, so behave as if one is the head and the
other the tail of a planckian rope for each pair.

This got the right results but is rather overcomplicated. Also, a simple
hashset would have been enough after all, but at least we could now
compute the intersections for fun.