10 cathode-ray tube
Simulate this nearly simplest possible CPU with a "virtual machine" that
tracks the x register as specified. This VM is slightly over-generic to
accommodate for future expansion as big elf corp strives for world
domination.

Track the clock in another loop inside the instruction loop; keep doing
clock iteration if the instruction needs it.

"During the cycle" really means before the instruction has done work.