10b the x register controls a sprite
The computer did not expand in this decade yet but the CPU interface
design to the outside world has now been finished. Produce a new pixel
for each clock cycle, and switch to a new scanline every 40 cycles. The
pixel is set if the x register is near enough the CRT cay for the pixel.