11 monkey in the middle
Play Keep Away as specified in the instructions, nothing fancy yet.
Except that annoying workaround to index the monkey array separately for
throwing and receiving monkeys. With a "for m in monkeys" loop, monkeys
would be owned already once with the loop's mut borrow, and
monkey[destination] couldn't be modified (except e.g. with refcell, but
repeatedly indexing is a pattern very much like refcell anyway).