11b keep worry levels manageable
With the worry level no longet getting divided by three, u64 would
overflow after just about 10 iterations. Bignum libraries could get too
slow even after 900 iterations or so based on research, so make use of
an interesting modular arithmetic phenomenon of the monkey behaviour.

The divisibility tests of each monkey are prime numbers. Not sure if
that counts, but for each monkey the resulting number would be
equivalent regarding the test with one that's the remainder from the
divisibility test. For this to work for each monkey, it's fine to
multiply all those test numbers together. If they were not primes, the
overall product could be optimized with least common multiple.

This does not affect part a of the puzzle, so take the modulus always
and hardcode the product of them all as the product of prime numbers
from 2 to 23; the input and the sample input are contained under these.
As the item worry level can still get multiplied by itself, the
remainder must stay smaller than sqrt(u64 max) = about u32 max =
4294967295. 2*3*..*19*23 = 223092870 is about 5% of that, so we're safe.
Any prime number product scan just under u32 max should also work.