12 hill climbing algorithm
It's adventure time on a heightmap. Solve the distance from start to end
with a simple breadth-first search on a graph that the heightmap
describes: each grid cell is a node, and neighbors are those that can be
accessed based on the elevation differences.

Spent too much time on a copypasta bug, it's not pos == end!