12b turn this into a hiking trail
Bruteforce :D each possible starting location of level a and take the
shortest route to E. The idea though is probably to do an inverse search
from E; that's just one BFS over the entire graph.

As a hack, return a huge value from bfs if the goal wasn't found rather
than panicing.