12: bfs backwards to avoid repeated searches
Run the search from the end node to the start in the first part,
carefully applying the edge rules in the opposite order. In the second
part, traverse the entire graph and keep track of the minimum distance
to any lowest point.

Now the bfs is quite a bit hardcoded for the exercise as the min is
always tracked and the search could be done only once for both puzzle
parts, but a bfs algorithm core isn't that complex anyway...