13 distress signal
Find which packets are in the right order based on the comparison
instructions. Keep the input as immutable strings rather than converting
to typed tree structures for an allocless recursive iteration exercise.
Recurse the two packets in pairs, comparing numbers or lists. The
recursion returns a ref to the end of the consumed data, so the caller
can continue.

As the "integers" in the list are from 0 to 10, preprocess the 10s in
the string into ':' that comes right after '9' in ascii. This makes
integer "parsing" a bit simpler of a task, as they're all single digit.

For the special case of one number vs a list, do the inverse of the
specification; still treat the number as an integer, but find the
corresponding integer (or empty list) in the list to compare against.
That is kind of the leftmost leaf node of the next subtree. This skips
over the entire list then so the next integer can be tested.

This entire business smells like a comparison of two trees and could
probably be much simpler with actual tree structures. At least stepping
the two together might have been a mistake.