13: make local "lists" for single-int comparison
Turns out the skipping of the other lists separately is not necessary.
It kind of was with no extra storage because the single integer could be
followed by more, but it's easy enough to store in a local tiny list of
constant size and then reuse the same comparison recursion.