13: add compare result type to ease the option mess
The returned l and r were only valid for None compare result (equal).
Make it explicit with types.

Sadly, "_ => return result" when "_" binds to only those choices of the
enum that bind no local variables could still return a ref to a local
according to the compiler. The world is not complete, so mirror those
returns annoyingly.