15b the distress signal is coming from nearby
If 4000000 is "nearby" then the tunnel maze must be huge. Refactor the
row search to use spans of min and max, as there is no need to look at
each coordinate one by one and there are so many of them. The spans are
merged as they're built to avoid duplicates. If the scan provides
exactly two separated spans that cover the x limits and their separation
leaves exactly one possible coordinate, that's our beacon.

This is still super hacky because the result is just printed in the
middle to reuse the code quickly.