17 pyroclastic flow
Rocks fall from the sky and pile up like tetris blocks played by hot
gas. Seems like a deception when the question is the height of the
rocks, and the movements repeat like that.

For now emulate this game just as-is, with each cell unit represented by
a coordinate in a hashset of rock fragments. This mission is probably
about finding a cycle in the heights. Right? Maybe the rows don't
explode like in tetris, maybe they do.