17b elephants are not impressed
The height increases start repeating a cycle after some time.
Interestingly the cycle is much shorter than the jet pattern length. Add
cycle detection to the loop, and once a cycle is found, use it rather
than the direct emulation.

Overengineer a signal cycle checker so it can perhaps be used later on
without reinventing again. This also simplifies the call site nicely.