16b one of the elephants to help
Extend the state node with the valve index that the big animal friend is
currently vacating. Add a new separate search to make use of it.

Now there are much more edges between search space nodes:

(stay, open, or move) x (elephant stays, elephant opens, elephant moves)

Otherwise the search is mostly the same as before.

The A*-like heuristic of not even trying an edge that could not possibly
lead to a better pressure result than the best so far seems to be
necessary to prune the choices enough to make this run even in that one
and half minutes, which still is very long; the search space should be
cut much more with some trick. It's perhaps not a great idea to hold the
time left in the search space node? Why did that even exist again?

One possible optimization could be to jump over the non-steaming valves
that are just tunnel parts, but time tracking is a bit tricky with that
if both of the travelers need to end up at some very specific nodes that
are different steps apart from current. The tunnel cells probably
contribute to a lot of back-and-forth walking. Maybe going to a tunnel
just for a u-turn is avoided with the visit scores, but just jumping
over the tunnel parts would make that worry unnecessary.

What if each traveler had a short-term plan to go to some particular
closed valve, rather than just trying what happens by visiting the
nearest neighbors? For now it's a lot of shuffling around for nothing.
In previous years' searches, "teleporting" right over to important nodes
by constructing a complete graph has caused success.