17: cache max height
The hashset gets rather big, and embarrassingly most of the time was
spent in figuring out the tallest rock fragment. Cache the max y.

Now it's nearly 50-50 cycle checker and the game. The game spends most
of the time in testing for overlap; an easy way to optimize would be to
bitmap each row and store the rows in a vec, but it's getting boring.