18b something seems off
The lava is indeed a hollow shell. Count only the surface faces that are
exposed to water and steam.

DFS the outer volume of the point set, which luckily is a nice shape, to
obtain a volume covering the outside. Then do similar face checks like
in the first part of the puzzle but count the cubes that touch the outer
volume, and don't care about other lava cubes.

The spherical structure could provide for some additional interesting
heuristics but that's not necessary.