19 not enough minerals
A collection of geodes is right there to be cracked open. Need robots to
crack them, of course. A robot needs material produced by other robots.
Looks much like year 2019 day 14 where fuel was produced from ore, but
this time the setting is a bit different with the time axis. Do a basic
DFS with some heuristics.

Prune the graph when there's no time left to possibly produce more
geodes than the maximum (probably not very efficient, as this equation
is so generous) and when the number of robots would become so large that
there would be no need for extra material produced.

Still takes a whole lot of time. Might be worth it to look into building
in inverse time, to start by requiring infinite geodes and then seeing
how the material for them could be achieved, and the material for that
recursively, like in 2019/14.