19: only build useful bot if possible
If there are resources to build a geode bot, it is always the best
option; prune the other search branches that would try something else in
that case. Similarly for obsidian bots as those are really needed for
geode bots.

BTW from runtime testing it seems the top level heuristic to prune if
best geodes can't be beat in remaining time is a really significant
factor.