20 grove positioning system
The grove coordinate file is encrypted. Run the specified linked list
stuff in brute-force manner; takes about 0.7 seconds. Might be better to
skip over links without flipping pairs like that, but off-by-ones in the
morning means NOT comfortable.

Also, moving left and right are interchangeable if the step count is
adjusted just right, but that's for later.