20b the values seem nonsensical
Apply the decryption key, of course. Multiply the values, and mix a bit
more. The mixing itself changes only slightly to support massive values
that wrap around the cyclic data. When a value hops forward, it's hopped
back to itself when the number of numbers minus one hops are done -
dangerous case for the fencepost error.