22 monkey map
Trace a path on a strangely-shaped board, sometimes colliding on solid
walls. This map, as most of them, fits best in a hashmap. If a position
is not in the map, it's void; if it's there, it's either open or a wall.

The world wraps around at the void. Easy enough.

I've done enough origami to see that the map looks like it'll probably
get folded into a 3d shape.