22b.0 input device isn't what the monkeys drew
The input device isn't that map but the faces of a cube folded flat, so
the wrapping rules are quite different.

How the edge pairs are mapped is not quite trivial to derive so the
mapping is hardcoded for now, and only done for the sample input in this
patch. The algorithm seems to be fine though:

For each open edge not immediately connected to another face in the map
data, encode the equivalent edge on the other face connected to it. The
equivalent edge is the face index (as x, y), how to turn the axis that
runs along the source edge, and where the origin of the axis lies in the
pair edge. The source edge axis is known from the direction the edge is
headed at; the axis direction is the same as with major axes, and the
origin is determined by that rule.

This brings the position to the right spot on the neighbor face on the
edge and also flips the heading with the same transform as the edge axis
vector is rotated.

It's probably possible to write these in more concise manner (such as "U
turn to the right") but for now the abstraction level is low.

Figuring out the edge pairs from the input data could be tricky, so
another plan for automatic edge discovery would be to think the cube as
a known fold-out shape with well defined edge warps and then to
transform the source faces into this "canonical form". To get the final
coordinate in map reality, the coordinate would be transformed back from
the virtual reality in the canonical form.