24 blizzard basin
This time the graph search happens such that time is a simple third axis
in the map; the blizzards move and are in different positions each
minute. Instead of moving the blizzards though, move the expedition
position in each compass direction to see if a collision would happen.
Fairly simple dijkstra/A*-ish otherwise: the priority queue score
optimizes distance to the goal and consumed minutes.

Maybe a basic BFS would work and be fast too as the network does not
branch out much. The minutes should probably be the first member in the
score, as now the first found goal is not the best one.