24b one elf forgot his snacks
Just run the same search multiple times between the start and goal
positions, keeping track of how much time has passed.

Bugfix: allow staying in the entry pos for longer. This was not very
clearly specified in the puzzle instructions.

This runs a bit long though, the iteration over all blizzards is
starting to show; the collision testing like that isn't really
hashmap-friendly.