24: sort the queue by minutes
The graph node space is 3D, trivially traversable by (x,y,time) and the
distance to the goal is only the x,y distance, but that does not take
into account the time axis. Put the consumed minutes as the first field
in the score, so the binary heap provides the quickest search faster.

The minutes still has to be stored in the node that is used as the visit
map key because time is part of the node space.