22: autodetect cube face connectivity information
All cube foldings are equivalent as far as edge wrapping goes. Assume a
"canonical" folding pattern and discover a mapping to translate the
input file (sample and puzzle input) into that representation. The
mapping is then used in another patch to perform the wrapping when
crossing edges on the path from the monkeys.

The map data stays as-is; the faces get "labels" in a canonical
coordinate system. The chosen canonical form is this one with two
symmetric flaps:

     F                                                 G
     E   or when the input is wider than it is tall: FEAD
    BAG                                                B
     D

The labels do not mean anything. One is G instead of C because its
rotation is easier to see on paper strips.

The mapping discovery assumes that at least one face is in a canonical
position; all those positions are tested and neighbors are processed
recursively when found.

This canonical folding includes the edge wrapping pairs including how
the edge axis rotates. The relative rotations are equivalent in all
possible folding patterns; this way the pairs need to be encoded only
once.