22: use autodetected face pairs to wrap
The autodetection provides face/face pairs at edges with edge rotation.
That's in the chosen canonical coordinate system, so translate map
coordinates into canonical coordinates, do the wrap, and translate back
from canon coords into map coords.

The rotation is still a bit clunky because it's done such that the edge
is a coordinate axis, but in reality it is may not be until doing
another transform to achieve that. The source edge axis is easy but the
destination for now hardcodes all 4*4 corners per rotation and edge.

The rotations are encoded with numbers of 90 degree turns that happen to
be composable as plain arithmetic all over the place; the edge rotation
describes how the face warps and also how the path facing (aka heading)
flips.

There are many ways to think how the face rotations happen; one is how
many 90-degree turns a face rotates, and one is which face index (0, 1,
2, 3 starting from the right) gets to be the new right edge.