22: flip the edge only at the end
With the logic structured just so, it becomes evident that the edge
offset does not contribute to the transforms and rotates each time
together with the heading. Take the src and dest heading difference and
use it to rotate the offset vector along the edge only once, after the
transforms have completed.