22: flatten destination corner map
The edge axis origin map has all the corners in clockwise order. With
this realization, delete the lookup table and substitute just the list
of the corners and an index table that tells where the origins are in
nominal non-rotated transforms:

   0     1
    +--->
    |   |
    |   |
    v   v
    ---->
   3     2

- When moving right, the dest edge has its origin at face top left
- When moving down, the dest edge has its origin at face top left
- When moving left, the dest edge has its origin at face top right
- When moving up, the dest edge has its origin at face top left

These then rotate as the wrap includes a rotation; for example, heading
up (3) and turning left (3) gets the origin at the corner 2, so the
source x axis becomes y axis pointing up.

This bears some similarity to the source axis corner selection that also
is the corners 0 0 1 3 but ordered as 1 3 0 0.