1b spelled out with letters
Some of the digits are not single digit bytes but actually text, so find
the first and last numeric or textual digits for the calibration value.
Some of those may overlap (like "twone"), so the non-overlapping
iterative matches in the regex crate are no use.

Building the regexes again for each test is massive bloat but the input
size is too small to worry about it.

Another clever trick could be to reverse the string and the searches,
but for now just consume everything and then find a match for the last
digit.