2 cube conundrum
Parse each game, going through those reveals of sets of cubes. Record
the highest need of cube colors at any time, and compare to the valid
count of 12, 13 and 14.

Looks like the separate sets could be useful in part 2, but for now it
would be sufficient to just look at the colors withour paying attention
to the sets.