4b win more scratchcards
The only way to win is to play more? Each match gets us more cards from
the list that follows the win. Instead of the given algorithm, count the
winning cards recursively. This does a whole lot of repetition and is as
stupid as doing fibonacci recursively, but the input size is small
enough to run in a second which is good enough for now and the code is
nice and readable like this.