6 wait for it
With the small play times, just try exhaustively each value from 1 to
playtime minus 1 and find how many ways there are.

Now with the beat test written down like that, it now looks obvious that
this is probably just a parabola with an easy solution on pen and paper.
But computer go brrrr.