7b a little more interesting
J is now a wildcard, becomes anything. Parse another set of cards for
where the joker card value is different, and run the jokered hand score
based on the normal rules so that jokers are replaced by the optimal way
to do it. Not too many scenarios to check manually.