10 pipe maze
Parse the peculiar ascii art for a graph of cells with two neighbors
each, and traverse the graph for the length of that single loop. The
farthest distance is half the loop length.

The graph is not easy to parse directly into a directional format so
there's no obvious next/prev pointers, just "pair A" and "pair B" so
keep track of the current heading to avoid getting stuck stepping back
and forth between two nodes when looping.