10b the animal never emerges
To find the area enclosed by the loop, walk through the loop again and
flood-fill the cells "to the right" of the walking direction. With a
clockwise walk this would be the inside area. It's not so clear yet
though so just do both and print a map; visually it is obvious which one
is the right answer.

Crossings per each scanline apparently could be a simpler algorithm,
similar to the generic "point inside triangle" where an odd number of
crossings would mean inside polygon and even number would be out.