10: autodetect clockwiseness
In this cell-based world the simplest loop has four turns, and a more
complex loop reduces to looking like four turns when the turns are
summed together. Walk all the pairs of move deltas in the loop and sum
the cross products; the cross product of two identical direction vectors
is 0, and for the 90 degree turns it's -1 or 1.

The "directionality" of a loop is unveiled by the sign of the sum of the
cross products. If the loop goes counterclockwise, reverse it. This way
the right-of cells must be inside the loop, so there's no need for a
human to look at a map.