12b the records were folded up this whole time
The search for part 1 is easily made faster and much more suitable for
longer sequences by just caching. Ah, the solution to all DP problems,
simply memoize it. The cache stores slices so this got a bit nasty with
explicit lifetimes. Some sort of generic "cache wrapper" could be
useful for AoCs in general, if possible.