14b a button on the side
A spin cycle rolls the rocks to each side repeatedly. Switch the
viewpoint and use the same rolling code, just rotate the map 90 degrees
and then slide. The spin count is a bit large, but it's reasonable to
expect that this spinning settles the rocks over time to a repeating
signal, so reuse a generic cycled signal detector from last years. The
detector apparently used integrals so it had to be adjusted a bit, and
now somehow just works.