17 clumsy crucible
There's a path to find in a grid-ish graph with weights on each grid
cell, and then the movement is constrained by limits on how to turn and
how much forward movement is possible. Use the constraints to build all
actual movement graph edges (left, right, forward 1/2/3) and dijkstra it
away to find the shortest path i.e. minimum heat loss.