19b sorting process still isn't fast enough
The rating ranges have reasonable limits, and turns out that the
workflows can be appied on ranges of numbers. This gets us subsets of
the entire range recursively, and in the end a whole range gets accepted
or rejected; the recursion sums up those leaf nodes. Fun but got to be
careful for off-by-ones with those lo/hi bounds.