21b the elf seems confused
The farm map repeats infinitely. By chance it's shaped such that the 2d
area repeats in a special pattern consistently enough. A simple
quadratic equation gives the number of plots reachable given the number
of map repetition cycles; the first plot counts to solve the quadratic
are easy enough to resolve in brute-force manner.

A more clever result probably exists from the cell parity and all (once
the map is "fully walked" the cells blink on and off) but this is enough
for now.