22 sand slabs
"Don't actually disintegrate any bricks - just determine what would
happen --" lol, hold my beer... Anyway, this is like 3D tetris. First,
play the tetris to the end so that all bricks are on the ground or on
top of each other, and then find redundant bricks and bricks on top of
the tower; those can be disintegrated.