23 a long walk
The map is a nice 2D grid maze. Process it with a simple DFS into a
smaller graph that keeps only the intersections, not tunnels. This
smaller graph is trivial to just search exhaustively for the longest
path, so do that. The slopes (directional parts) act as walls when
DFSing further, but they still need to be considered when building the
intersection nodes. Turns out that the slopes also make the graph
cycle-free, so the search for longest path and the puzzle requirement to
never visit a spot twice becomes proven by the construction of input.