24b throw it just right
Use the hint from part 1 and find the solution in the xy plane first.
Key insight is that when the rock is happily flying about at the right
speed, from its point of view all the hailstones look like they will
collide. Assume that the magical speed is reasonable and try x and y
velocities from 1 to 300. Then in this velocity-adjusted hailstorm the
collision time is any intersection of two hailstones. The time and line
equations give the full 3D solution from an arbitrary hailstone pair.

Brute-force trying those velocities lasts an embarrassing 45 seconds for
my input (x 99, y 240). Some clever culling could be in order.