4 ceres search
Starting at each X in the map, try if it leads to a MAS in any straight
or diagonal direction.

Lost some time in this by misreading the instructions like so that the
xmas pattern could also twist, hence the overcomplicated recursion.

For these 2d grid maps, some aoc library to provide indexing on i32/i64
and/or sentinel values at the edges could make the "as usize" less bad.
Maybe an investment to do on some day.