5 print queue
Does each day come with a reference to a previous year? So it seems with
the history theme. Anyway, for each rule, run a silly O(update size^2)
test to see if the rule's pair exists the wrong way. Could also track
the first seen number of the pair and do the test on the second, if any.