7b well-hidden elephants
Implement the new concatenation operator and just add it to the search;
carrying the "this is part b" flag along is not pretty but is good
enough for today. Certainly the limited selection of operators is chosen
to give a hint to short-circuit the search early as the number can only
grow (and become too big), but we're still fast enough.