Optimize 6 for under 0.4 seconds
First, only consider the positions the guard actually visits; this
already cuts the search space to about 30% of the map (~two seconds).
The key trick though is to record the path of guard poses and loop it
backwards, trying each position for the obstruction and simulating the
route again from the pose just prior to that one. Looping backwards
allows to iteratively clear the visit pose map so that it matches the
path until that point and can be reused.

However, there's one corner case: some positions are visited more than
once because the walk paths can cross. If a position is visited earlier
with another heading, don't try obstructing it and simulating later
because that later would not happen anymore.

(The vec^2 for the map is becoming real annoying. Gotta make a toolkit.)