9b move whole files instead
Don't expand the disk but keep a list of data descriptors representing
whole files or contiguous empty blocks. Iterate the files in
back-to-front order and try moving them, substituting empty for moved
ones and squashing new empty data with its surroundings. A destination
descriptor may split if the file doesn't fill it exactly.

To speed things up and avoid looking up the next file from the front,
keep an index to the rightmost pos and start looking from there.

This gets the lengthy off-by-1 debug prize of the year so far with that
disk.iter().take(src_idx) that was src_idx-1.