10 hoof it
Run a trivial DFS on the topographic map, storing the visited positions
to avoid stepping in the same place twice and counting extra paths. The
search ends at a node of score 9, and each search returns the number of
high grounds found.

The vec^2 is getting real boring with usize indexing, so start crafting
a map type to work with signed coordinates. The use ergonomics appears
to be about the same as hashmap.