10b a second way to measure a trailhead
Implement the rating calculation; it's exactly the same as for part 1
with just the visit map removed. This gets the number of paths instead
of the number of reached high spots.

That copypasta should be cleaned up at once, and even got into some
trouble by accidentally calling search instead of search2.