15 warehouse woes
This yet another 2D grid simulation is about pushing boxes around until
they hit a wall. A row of boxes moves one step if and only if there's
empty space after it. The boxes all look the same, so moving a row is
equivalent to moving just the head into the empty space.