15b a second warehouse
Everything is twice as wide. Now the boxes are one size in one dimension
and two sizes in another dimension. They push each other horizontally
just like in the first part except that an entire row has to move
because the left and right sides don't look the same so the "just move
the head" optimization goes away.

Vertical movement is more complex as one box can move two boxes. The
push cycle becomes not just a tree but a graph, so BFS the graph to find
all boxes. If there's a wall, the BFS cuts short. At the end of a
successful search, move all boxes in order of last seen first, and the
nearest to the robot last. This graph business actually would work for
horizontal moves as well but that would need generalization over axes.