16 reindeer maze
Need to find the shortest path in a map, so dijkstra it is. Graph nodes
are (position, heading) pairs; edges go into neighbor cells (weight 1)
and in-place left/right turns (weight 1000). The best path is the
shortest one from start pose to any heading at the end position.

Should really start using some use'able toolkit to put that map in.