16b the best spot to sit
Extend the dijkstra search to collect backward edges so that traveling
from the goal is easy. To have all possible paths, also try nodes that
have the same cost as already known, not just better. With the edge and
distance maps, collect all tiles on the path from the end until no edges
remain and we're at the start node. The number of tiles is the number of
possibly best spots.

The search leaves some funny neighbors in the edge map so not all edges
can be taken, maybe that's a leftover detour later overwritten by a
better path or something. If so, maybe the dijkstra should be made to
prune such edges away.