17b a seems to have been corrupted
Find what value for register A outputs a copy of the program itself. The
program appears to output a value for every three bits in the register,
halting when the register goes to 0, so build it in a recursive search
in pieces of three bits.

Those output values however are affected by more than those three bits.
That division appears to be constructed such that after 10 bits the
output wraps again, and this could be true for all input programs; maybe
the difference is just the magic constants that mix the randomness.

TODO: run the search from the biggest digit first to loop just to 8, not
1024. But can the biggest digit be a special case needing more?