18 ram run
With the fallen bytes in a simple two-dimensional grid, run a shortest
path algorithm to get from the top left corner to the bottom right
corner after 1024 steps. The pathfinding is a year old copypasted
vintage Dijkstra, but the grid is so consistent in costs that a plain
BFS would also work.