18b not used to moving around
Determine the first byte that will cut off the path. The map is small
enough so that is trivial brute force work to complete in three seconds.
The only special trick is to cut the search short instead of looking up
each possible node which wasn't really necessary even earlier. Three
seconds is faster than debugging off-by-ones in a binary search that
this shall be later optimized into.