19 linen layout
We have an infinite supply of towels with various patterns of colors and
a list of desired sequences of colors. Count how many desired sequences
are possible using the set of patterns: try greedily each towel pattern
at the beginning of the sequence, and recurse with the remaining
sequence for those that match a towel, until there is no pattern
remaining. Memoize this with the pattern string as key.