20 race condition
Now we have a map that, after reading the spec a hundred times, turns
out to not be a maze at all but a straight line. It may become not a
straight line though after applying a cheat: replace two neighboring
cells with empty space. While thinking for a better way, do silly
copypasta based brute force: try cheating each cell pair in the map,
find the distance from start to finish, and track those that save 100
steps or more.

At almost three minutes of runtime, this is one of the most embarrassing
dumb solutions so far.