20b programs seem perplexed
Try cheats of at most 20 picoseconds, not just two. After some silly
attempts at trying to apply the cheat in dijkstra code to make it aware
of those warps through walls, take the distance map from the initial
search and use it to see how many steps would be saved if a direct
manhattan distance would be taken for any two points along the road.

This is still only slightly adapted from the first part; it would be
better to try the nodes along the path found in the search instead of
the map cells which is wasteful and suffers from seeing each path twice
as mirrored pairs. 2.5 seconds for both parts is not too bad though from
the previous multi-minute hack.